Moved testPositionRespectsTimeout to a more suitable location in PlaintextConsumerTest
